[정인호]Vo 패키지 변경 통합작업
