+++ src/main/java/com/tree/gdhealth/branch/api/BranchApiMapper.java
@@ -1,6 +1,6 @@
 package com.tree.gdhealth.branch.api;
 
-import com.tree.gdhealth.branch.vo.Branch;
+import com.tree.gdhealth.vo.Branch;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
