+++ src/main/java/com/tree/gdhealth/employee/api/EmployeeApiMapper.java
@@ -1,6 +1,7 @@
 package com.tree.gdhealth.employee.api;
 
-import com.tree.gdhealth.headoffice.emp.vo.Employee;
+import com.tree.gdhealth.vo.Employee;
+import com.tree.gdhealth.vo.EmployeeDetail;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
