+++ src/main/java/com/tree/gdhealth/headoffice/emp/EmpMapper.xml
@@ -54,7 +54,7 @@
 		WHERE e.employee_id = #{employeeId}
 	</select>
 	
-	<insert id="insertEmployee" parameterType="com.tree.gdhealth.headoffice.emp.vo.Employee">	
+	<insert id="insertEmployee" parameterType="com.tree.gdhealth.vo.Employee">
 		INSERT INTO employee(
 			branch_no,
 			employee_id,
@@ -68,7 +68,7 @@
       		  SELECT LAST_INSERT_ID()
    		</selectKey> 
 	</insert>
-	<insert id="insertEmployeeDetail" parameterType="com.tree.gdhealth.headoffice.emp.vo.EmployeeDetail">
+	<insert id="insertEmployeeDetail" parameterType="com.tree.gdhealth.vo.EmployeeDetail">
 		INSERT INTO employee_detail(
 			employee_no,
 			employee_name,
@@ -83,7 +83,7 @@
 			#{employeeGender}
 		)
 	</insert>
-	<insert id="insertEmployeeImg" parameterType="com.tree.gdhealth.headoffice.emp.vo.EmployeeImg">
+	<insert id="insertEmployeeImg" parameterType="com.tree.gdhealth.vo.EmployeeImg">
 		INSERT INTO employee_img(
 			employee_no,
 			employee_img_origin_name,
