+++ src/main/java/com/tree/gdhealth/mybatis/typehandler/OrderStatusTypeHandler.java
@@ -1,5 +1,6 @@
-package com.tree.gdhealth.sportsequipment.vo;
+package com.tree.gdhealth.mybatis.typehandler;
 
+import com.tree.gdhealth.sportsequipment.vo.OrderStatus;
 import org.apache.ibatis.type.JdbcType;
 import org.apache.ibatis.type.MappedTypes;
 import org.apache.ibatis.type.TypeHandler;
