+++ src/main/resources/mybatis-config.xml
@@ -3,8 +3,8 @@
 <configuration>
     <!-- Your MyBatis configuration settings go here -->
     <typeHandlers>
-        <typeHandler handler="com.tree.gdhealth.sportsequipment.vo.OrderStatusTypeHandler"/>
-        <typeHandler handler="com.tree.gdhealth.sportsequipment.vo.ImgTypeTypeHandler"/>
+        <typeHandler handler="com.tree.gdhealth.mybatis.typehandler.OrderStatusTypeHandler"/>
+        <typeHandler handler="com.tree.gdhealth.mybatis.typehandler.ImgTypeTypeHandler"/>
     </typeHandlers>
     <!-- Type aliases, mappers, etc. -->
 </configuration>
