+++ src/test/java/com/tree/gdhealth/sportsequipment/vo/SportEquipmentMapperTest.java
@@ -1,5 +1,8 @@
 package com.tree.gdhealth.sportsequipment.vo;
 
+import com.tree.gdhealth.vo.SportsEquipment;
+import com.tree.gdhealth.vo.SportsEquipmentImg;
+import com.tree.gdhealth.vo.SportsEquipmentOrder;
 import org.apache.ibatis.annotations.Mapper;
 
 /**
